Customize defaults in models
In model itself. Still needs to allow using a different primary key name

+++ db/migrations/20170127143151_create_sign_in_credentials.cr
@@ -1,9 +1,9 @@
 class CreateSignInCredentials::V20170127143151 < Avram::Migrator::Migration::V1
   def migrate
     create :sign_in_credentials do
-      primary_key id : Int32
+      primary_key id : Int64
       add_timestamps
-      add user_id : Int32
+      add user_id : Int64
     end
   end
