+++ db/migrations/20180113081408_create_companies.cr
@@ -1,7 +1,7 @@
 class CreateCompanies::V20180113081408 < Avram::Migrator::Migration::V1
   def migrate
     create :companies do
-      primary_key id : Int32
+      primary_key id : Int64
       add_timestamps
       add sales : Int64
       add earnings : Float
