+++ db/migrations/20180802180356_create_blobs.cr
@@ -1,9 +1,9 @@
 class CreateBlobs::V20180802180356 < Avram::Migrator::Migration::V1
   def migrate
     create :blobs do
-      primary_key id : Int32
+      primary_key id : Int64
       add_timestamps
-      add doc : JSON::Any?, default: JSON::Any.new({ "defa'ult" => JSON::Any.new("val'ue") })
+      add doc : JSON::Any?, default: JSON::Any.new({"defa'ult" => JSON::Any.new("val'ue")})
     end
   end
